+++ src/components/ContentLists/index.js
@@ -1,6 +1,0 @@
-import './ContentLists.css';
-
-import { SongList } from './SongList';
-import { VideoList } from './VideoList';
-
-export { SongList, VideoList };