+++ src/components/Lists/SearchList.js
@@ -12,7 +12,7 @@ export const SearchList = ({ values, type, location, className }) => {
       } else if (value.collectionType === 'Album') {
         return <AlbumItem key={index} value={value} />;
       } else if (value.kind === 'music-video') {
-        return <VideoItem key={index} value={value} />;
+        return <VideoItem key={index} value={value} contentList={false} />;
       }
 
       return false;
@@ -29,9 +29,7 @@ export const SearchList = ({ values, type, location, className }) => {
           </Link>
         </div>
 
-        <div className="scrollable">
-          <div className={`container--horizontal ${className}`}>{data}</div>
-        </div>
+        <div className={`container--horizontal ${className}`}>{data}</div>
       </div>
     );
   } else {